- Basic wandb capture of the training and val cycles. - Focal Loss implementation (it works! But does it... "work"?) - Logging of model shape to a separate csv. - Minor bits of refactoring to structure the code a bit better.
If you want to use Focal Loss I have altered the param block in the yml. Not supplying this block will default to BCE loss instead.
Wandb is configuration "safe" in that if you don't supply credentials it will ignore the wandb logging without throwing errors. At the moment the default step metric is epochs.