Added a test saved model and kaggle output function.
Note that heavy refactoring has taken place to main.py to allow for a clean separation of a MelanomaTrainer and a new MelanomaTest class. The refactoring has brought the train, validate and train_batch methods out of the instantiated class. Mainly to allow them to be called independantly of the classes themselves.